+++ workspace/apps/perpetuals/contracts/src/core/components/assets/interface.cairo
@@ -21,7 +21,7 @@ pub trait IAssets<TContractState> {
     fn add_synthetic_asset(
         ref self: TContractState,
         asset_id: AssetId,
-        risk_factor_tiers: Span<u8>,
+        risk_factor_tiers: Span<u16>,
         risk_factor_first_tier_boundary: u128,
         risk_factor_tier_size: u128,
         quorum: u8,
